+++ src/airflow_export_to_object_store/unload/__init__.py
@@ -3,10 +3,16 @@
 from __future__ import annotations
 
 from .base import UnloadStrategy
+from .bigquery import BigQueryUnloadOptions, BigQueryUnloadStrategy
+from .redshift import RedshiftUnloadOptions, RedshiftUnloadStrategy
 from .snowflake import SnowflakeUnloadOptions, SnowflakeUnloadStrategy
 
 __all__ = [
     "UnloadStrategy",
+    "BigQueryUnloadOptions",
+    "BigQueryUnloadStrategy",
+    "RedshiftUnloadOptions",
+    "RedshiftUnloadStrategy",
     "SnowflakeUnloadOptions",
     "SnowflakeUnloadStrategy",
 ]